Reorganized and implemented RelItem
