+++ src/headers/file.rs
@@ -1,6 +1,5 @@
-use crate::errors::{Error,Result};
-use crate::common::{Width,Layout};
-use crate::common::{Item,ranges::*};
+use crate::errors::{Result};
+use crate::common::{Width,Layout,Item,ranges::*};
 
 /// The ELF file header parsed from the beginning of the file
 #[derive(Debug)]
