+++ src/tables/items/mod.rs
@@ -1,0 +1,13 @@
+
+mod traits;
+mod string;
+mod symbol;
+mod relocation;
+
+pub use traits::TableItem;
+pub use string::StringItem;
+pub use symbol::SymbolItem;
+pub use relocation::{
+    RelItem,
+    RelaItem
+};
