+++ src/tables/tables/mod.rs
@@ -1,0 +1,13 @@
+
+mod relocation;
+mod string;
+mod symbol;
+mod table;
+
+pub use table::Table;
+pub use string::StringTable;
+pub use symbol::SymbolTable;
+pub use relocation::{
+    RelTable,
+    RelaTable
+};
